+++ CHANGELOG.md
@@ -2,5 +2,8 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Horatum/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Horatum/releases)
 
-- Added: The boss combat encouters in mythic raids (flexible) are now being tracked
-- Added: TOC version for patch 12.0.7 [retail]
+**Important note: All settings have been reset to default values.**
+
+- Added: Profile settings with account profile and character profile selection
+- Updated: deDE, enUS localization
+- Minor code adjustments